requirements.txt
# arrow==1.4.0
# asgiref==3.10.0
# binaryornot==0.4.4
# certifi==2025.10.5
# chardet==5.2.0
# charset-normalizer==3.4.4
# click==8.3.0
# cookiecutter==2.6.0
# Django>=4.2,<5.0
# django-browser-reload==1.21.0
# django-tailwind==4.2.0
# djangorestframework==3.16.1
# djangorestframework_simplejwt==5.5.1
# djongo==1.3.7
# dnspython==2.8.0
# gunicorn==23.0.0
# idna==3.11
# Jinja2==3.1.6
# markdown-it-py==4.0.0
# MarkupSafe==3.0.3
# mdurl==0.1.2
# mongoengine==0.29.1
# packaging==25.0
# Pygments==2.19.2
# PyJWT==2.10.1
# pymongo==3.11.4
# python-dateutil==2.9.0.post0
# python-dotenv==1.1.1
# python-slugify==8.0.4
# pytz==2025.2
# PyYAML==6.0.3
# requests==2.32.5
# rich==14.2.0
# setuptools==80.9.0
# six==1.17.0
# sqlparse==0.2.4
# text-unidecode==1.3
# tzdata==2025.2
# urllib3==2.5.0
# whitenoise==6.11.0


Django>=4.2,<5.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
djongo==1.3.7
pymongo==4.9.2
python-dotenv==1.1.1
gunicorn==23.0.0
whitenoise==6.11.0
django-tailwind==4.2.0
django-browser-reload==1.21.0
